notification_websocket_service/requirements.txt
fastapi==0.111.0
pydantic-settings==2.2.1
websockets==12.0
aiohttp==3.9.5
aio-pika==9.4.1
pydantic==2.7.1
pyjwt==2.8.0
uvicorn==0.29.0
python-jose[cryptography]==3.3.0
aiologger==0.7.0
tenacity==8.2.3